requirements.txt
requests>=2.31.0
urllib3>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
pydantic>=2.0.0
python-dateutil>=2.8.0
cryptography>=41.0.0
pyjwt>=2.8.0
colorama>=0.4.0
tqdm>=4.65.0
jsonschema>=4.19.0
markdown>=3.4.0
xlsxwriter>=3.1.0